+++ functions.py
@@ -6,12 +6,17 @@ def minus(a, b):
     """Subtract b from a."""
     return a - b
 
+def share(a, b):
+  """ """"
+    return a // b
+  
 def modulo(a, b):
     """a mod b"""
     if b == 0:
         print("Can't operate!!")
     else:
         return a % b
+      
 def multiply(a, b):
     """Multiply two numbers"""
     return a * b